+++ pddl/strips_ground_datalog.h
@@ -36,6 +36,15 @@ int pddlStripsGroundDatalog(pddl_strips_t *strips,
                             const pddl_ground_config_t *cfg,
                             pddl_err_t *err);
 
+/**
+ * Ground PDDL into STRIPS using the Gringo grounder from the external
+ * libraly clingo that must be compiled-in.
+ */
+int pddlStripsGroundGringo(pddl_strips_t *strips,
+                           const pddl_t *pddl,
+                           const pddl_ground_config_t *cfg,
+                           pddl_err_t *err);
+
 #ifdef __cplusplus
 } /* extern "C" */
 #endif /* __cplusplus */